ConfigureHostGroups task refactoring
Pulls out ConfigureHostGroups so it runs indpependently from
ManageContent so exceptions can be handled specific to each task
