+++ server/app/lib/actions/fusor/content/manage_content.rb
@@ -14,6 +14,7 @@ module Actions
   module Fusor
     module Content
       class ManageContent < Actions::Fusor::FusorBaseAction
+        include ::Fusor::RepoHelpers
 
         def humanized_name
           _("Manage Content")
@@ -29,15 +30,11 @@ module Actions
           sequence do
             content = SETTINGS[:fusor][:content]
             host_groups = SETTINGS[:fusor][:host_groups]
-            product_map = {rhev: [:rhevm, :rhevh, :rhevsh],
-                           cfme: [:cloudforms],
-                           openstack: [:openstack],
-                           openshift: [:openshift]
-            }
 
+            # TODO: Move to host groups?
             enable_smart_class_parameter_overrides
 
-            product_content_details = product_map.each_with_object([]) do |(deploy, product_types), details|
+            product_content_details = PRODUCT_MAP.each_with_object([]) do |(deploy, product_types), details|
               details << product_types.map { |type| content[type] } if deployment.deploy?(deploy)
             end
             product_content_details = product_content_details.flatten(2).uniq.compact
@@ -55,18 +52,6 @@ module Actions
             plan_action(::Actions::Fusor::Content::PublishContentView,
                         deployment,
                         yum_repositories(repositories)) if deployment.lifecycle_environment_id
-
-            product_map.keys.each do |deploy|
-              hostgroup = host_groups[deploy]
-
-              if deployment.deploy?(deploy) && hostgroup
-                configure_activation_keys(deployment, hostgroup)
-                plan_action(::Actions::Fusor::ConfigureHostGroups,
-                            deployment,
-                            deploy,
-                            hostgroup)
-              end
-            end
           end
         end
 
@@ -83,63 +68,6 @@ module Actions
               end
             end
           end
-        end
-
-        def yum_repositories(repositories)
-          repositories.select { |repo| repo.content_type == ::Katello::Repository::YUM_TYPE }
-        end
-
-        def retrieve_deployment_repositories(organization, product_content)
-          repos = []
-          product_content.each { |details| repos << find_repository(organization, details) }
-          repos
-        end
-
-        def configure_activation_keys(deployment, hostgroup)
-          content = SETTINGS[:fusor][:content]
-
-          hostgroup[:host_groups].each do |subgroup|
-            next unless subgroup[:activation_key]
-
-            content_key = subgroup[:activation_key][:content]
-            product_content = content[content_key]
-
-            # find the subset of repos that belong to this activation key
-            repositories = retrieve_deployment_repositories(deployment.organization, product_content)
-            repositories = yum_repositories(repositories)
-
-            plan_action(::Actions::Fusor::ActivationKey::ConfigureActivationKey,
-                        deployment,
-                        subgroup,
-                        repositories
-                       )
-          end
-        end
-
-        def find_repository(organization, repo_details)
-          product = ::Katello::Product.where(:organization_id => organization.id,
-                                             :cp_id => repo_details[:product_id]).first
-
-          if product
-            product_content = product.productContent.find do |content|
-              content.content.id == repo_details[:repository_set_id]
-            end
-
-            substitutions = { basearch: repo_details[:basearch], releasever: repo_details[:releasever] }
-            unless repository = repository_mapper(product, product_content.content, substitutions).find_repository
-              fail _("Unable to locate repository for: Product '%{product_name}',"\
-                   " Repository Set '%{repo_set_name}'") %
-                       { :product_name => product.name, :repo_set_name => product_content.content.name }
-            end
-          else
-            fail _("Product '%{product_name}' does not exist. Confirm that a manifest"\
-                 " containing it has been imported.") % { :product_name => repo_details[:product_name] }
-          end
-          repository
-        end
-
-        def repository_mapper(product, content, substitutions)
-          ::Katello::Candlepin::Content::RepositoryMapper.new(product, content, substitutions)
         end
 
       end
